Suchir/LLM Agent/requirements.txt
python-dotenv>=0.19.0
pydantic>=2.0.0
langchain>=0.1.0
autogen>=0.2.0
python-json-logger>=2.0.0
transformers>=4.36.0
torch==2.1.0
torchvision==0.16.0
huggingface-hub>=0.19.0
ijson>=3.2.3 